test: split CLI error coverage
